# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Check the status of the REST API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/xtz/delegations": {
            "get": {
                "description": "List all the aggregated new delegations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "delegations"
                ],
                "summary": "List Delegations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter results by year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DelegationResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.SimpleErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.DelegationResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "block": {
                    "type": "string"
                },
                "delegator": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "errorhandling.SimpleErrorMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}